+++ RELEASE_NOTES.md
@@ -22,6 +22,16 @@ RELEASE NOTES
 =============
 
 
+3.99.0 - March ??, 2026
+=======================
+
+Tools
+-----
+
+* ADDED: Support to the `logtalk_tester.sh` and `logtalk_tester.ps1` testing
+automation scripts for concurrent running of tests sets with a new `-j` option.
+
+
 3.98.0 - February 27, 2026
 ==========================